+++ packages/db/src/index.ts
@@ -1,6 +1,7 @@
 export * from "./client.js";
 export * from "./schema.js";
 export * from "./repositories.js";
+export * from "./rls.js";
 export * from "./filters.js";
 export * from "./jobs.js";
 export * from "./hubspot.js";